Change in folder structure Insert SearchBar component inside Store(Products) folder, split components inside this same folder into different nested folders, move NotFound component to Misc folder and create it's own nested folder
