+++ client/src/components/Store (Products)/SearchBar/_SearchBar.scss
