+++ client/src/components/Store (Products)/Store.jsx
@@ -1,13 +1,13 @@
 import React, { useState } from "react";
-import SearchBar from "../SearchBar";
-import ProductCards from "./ProductCards";
+import SearchBar from "./SearchBar/SearchBar";
+import ProductContainer from "./ProductContainer/ProductContainer";
 
 export default function Store() {
   const [query, setQuery] = useState({ by: "product" });
   return (
     <div className="sectionContainer">
       <SearchBar setQuery={setQuery} query={query} />
-      <ProductCards query={query} />
+      <ProductContainer query={query} />
     </div>
   );
 }
